Check possibilities for impossibilities
Looks over the possibilities and sees if they would not let the black
squares connect. There is still more to be done for that (e.g. checking
if a possibility leaves a 2x2 behind) but it solves some more cases now.

There's also a bunch of code in here I don't remember writing or know
what it does. I'll have to figure that out later.